PDF_Scripts/requirements.txt
pdfminer-six
openai
langchain
tiktoken
pinecone-client 
sentence_transformers
